Refactor naming convention from plural to singular
Change the name of the component from Replacements to ReplacementChange the name of the controller from replacementsController to replacementControllerChange the name of the model from replacementsModel to replacementModelChange the name of the router from replacementsRouter to replacementRouterChange the name of the file from replacements.js to replacement.jsUpdate the references and imports accordingly
